Add cascade delete for user entity
Added the feature of detaching all the phones from the person entity when that
person is deleted. Also, a phone entity cannot be deleted when any person
is assigned to it and throws a constraint violation exception as expected
per the project requirements. However, still not able to catch that
exception in order to return a HTTP 400 error (as per the project
requirements).